refactor: reorganise gdal helper
