+++ scripts/stac/imagery/create_stac.py
@@ -4,7 +4,8 @@ from linz_logger import get_log
 
 from scripts.files.files_helper import get_file_name_from_path
 from scripts.files.geotiff import get_extents
-from scripts.gdal.gdalinfo import GdalInfo, gdal_info
+from scripts.gdal.gdal_helper import gdal_info
+from scripts.gdal.gdalinfo import GdalInfo
 from scripts.stac.imagery.item import ImageryItem
